# auth_service/openapi.yaml
openapi: 3.0.3

info:
  title: Gratia Authentication Service API
  version: 1.0.0
  description: |
    Authentication and authorization service for the Gratia platform.

    Features:
    - User registration
    - JWT authentication
    - Refresh token rotation
    - Session management
    - Password reset workflows
    - Protected user endpoints
  contact:
    name: Aditya Waradkar
    email: aditya@example.com

servers:
  - url: http://localhost:8080
    description: Local Development Server

tags:
  - name: Authentication
    description: Identity and token lifecycle management
  - name: User
    description: Authenticated profile operations
  - name: Password
    description: Secure account recovery flows
  - name: Monitoring
    description: Infrastructure health checks

security:
  - BearerAuth: []

paths:
  /register:
    post:
      summary: Register a new user
      description: Creates a new authenticated user account with strict role validation.
      operationId: registerUser
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account provisioned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Malformed payload or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email address already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /login:
    post:
      summary: Authenticate user
      description: Verifies credentials and provisions stateful and stateless tokens.
      operationId: loginUser
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session established successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid authentication credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /refresh:
    post:
      summary: Rotate authentication tokens
      description: Consumes a valid refresh token to issue a new credential pair.
      operationId: refreshTokens
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token rotation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Token expired, revoked, or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /logout:
    post:
      summary: Terminate session
      description: Permanently revokes the provided refresh token.
      operationId: logoutUser
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Session successfully destroyed
        '401':
          description: Invalid token provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /me:
    get:
      summary: Retrieve authenticated profile
      description: Returns the user details associated with the current JWT.
      operationId: getCurrentUser
      tags:
        - User
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User record no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /forgot-password:
    post:
      summary: Initiate account recovery
      description: Generates a secure reset token if the email exists.
      operationId: forgotPassword
      security: []
      tags:
        - Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Recovery process initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
        '400':
          description: Malformed email payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /reset-password:
    post:
      summary: Replace credentials
      description: Overwrites user password using a verified recovery token.
      operationId: resetPassword
      security: []
      tags:
        - Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Credentials successfully updated
        '400':
          description: Invalid token or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Container readiness probe
      operationId: healthCheck
      security: []
      tags:
        - Monitoring
      responses:
        '200':
          description: Service is accepting traffic
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: auth_service

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass123!
        role:
          type: string
          enum:
            - DONOR
            - NGO
            - ADMIN
            - USER
          default: USER

    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    RefreshRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8

    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - DONOR
            - NGO
            - ADMIN
            - USER
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PasswordResetResponse:
      type: object
      properties:
        resetToken:
          type: string
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string